# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moment, 
    thanks to WASAPhoto! You can upload your photos directly from your 
    PC, and they will be visible to everyone following you.
  version: 1.0.0
...
paths:
...
  /photo/{photoId}/likes/{username}:
    parameters:
      - $ref: "#/components/parameters/photoId"
      - $ref: "#/components/parameters/username"
    put:
      tags: ["likes"]
      summary: like the photo with the id
      description: like the photo with the id with the authenticated user
      operationId: likePhoto
      responses:
        '201':
          description: Sucessfuly liked photo
        '400':
          description: Invalid parameters supplied
        '404': 
          description: Photo not found
      security:
      - bearerAuth: [] 
    delete:
      tags: ["likes"]
      summary: unlike the photo with the id
      description: remove the aunthenticated user's like from the photo with the id
      operationId: unlikePhoto
      responses:
        '204':
          description: Sucessfuly unliked photo
        '400':
          description: Invalid parameters supplied
        '403':
          description: Forbidden action
        '404': 
          description: Photo not found
      security:
      - bearerAuth: [] 
components:
...
  parameters:
    username:
      schema:
        description: Valid username
        type: string
        example: Maria
        pattern: 'ˆ.*?$'
        minLength: 3
        maxLength: 16
      name: username
      in: path
      description: The name that needs to be fetched
      required: true
    photoId:
      schema:
        description: Identifier of the photo
        type: integer
        example: 1
      name: photoId
      in: path
      required: true
      description: Photo Id
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer